morpheus-metrics-dashboard/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
asttokens==2.4.1
attrs==24.2.0
bitarray==2.9.2
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.0
click==8.1.7
cytoolz==0.12.3
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
dune_client==1.7.5
eth-account==0.13.3
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
executing==2.1.0
fastapi==0.104.1
fastapi-utils==0.7.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
marshmallow==3.22.0
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
ndjson==0.3.1
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycryptodome==3.20.0
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
Pygments==2.18.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.27.0
toolz==0.12.1
traitlets==5.14.3
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240821
types-PyYAML==6.0.12.20240808
types-requests==2.32.0.20240712
types-setuptools==74.0.0.20240831
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.23.2
wcwidth==0.2.13
web3==7.0.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.9.4
